WIP - define new module and move around some classes and objects
