+++ modules/core/shared/src/main/scala/io/circe/package.scala
@@ -1,3 +1,5 @@
 package io
 
-package object circe
+package object circe {
+  val Printer = io.circe.prin
+}
